Initial support for CDB on JSON files.
This is a preliminary commit of the implementation that flushes cdb rdb
pojos in a set of JSON files. It does not yet do what we need. In
particular:
- while RDB pojos includes sets of other objects (for example the
supervisor has a set of DASU that in turn has a set of ASCE) we want to
have in pojos only the IDs otherwise the files are too complex
- this version only stores one IASIO in ISIO/IASIO.json instead of an
array of them
- reading from JSON files into java pojos is not yet implemented

Use Jakson2 for reading/writing JSON files.