Write JSON files without inclusion but with object IDs.
Jakson2 serialization works pretty but the JSON files it produces
includes all the linked objects (for example the ASCEs that runs inside
a DASU) making the filee very long, unreadable and almost impossible to
edit.
This version replaces object inclusion with IDs (for example the IDs of
the ASCE that run inside a DASU) using a set of pojos derived from RDB
pojos.

Note that object inclusion is not possible also because it is recursive
and cause an OOM exception.