Includes resource files to the jar.
Adds files in the resources folder to the jar built compiling java and
scala sources.
With this version only one jar is created for java and scala classes and
the resources. scalaJarName and javaJarName are deprecated and jarName
must be used.

The resources in the jar preserve the path they have in the resource
folder.