Thread factory for the plugin.
The thread factory creates thread for the plugin with common setup. All
threads belong to the same thread group and are daemons.
At this stage I am not sure that plain threads are the concurrency
mechanism to use but still makes sense to have this class for the
developers of plugins.