The filter needs the ID of the monitor point produced as this is what will be sent to the core of the IAS.
Some other refactoring.

Committing a lot of changes mostly being afraid to loose everything by
mistake. This version of the plugin contains the skeleton of the tool
but a lot of things still need to be completed.
Tests pass but they test only the filtering that has not been changed
this time.