Reads the configuration asynchronously.
PluginConfigDao is the interface to implement by the readers of the
configuration. 
The getPluginConfig() method returns a Future to allow async. reading
especially useful for slow readers like those reading and parsing JSON
files or accessing a remote dabase.
The PluginConfigFileReader async. reads the configuration from a JSON
file.
Test updated accrodingly.