Returns the number of messages sent to the core of the IAS.
numOfMessagesSent() returns the number of messages ent to the core of
the IAS between 2 invocations of the same method. The purpose is to
allow the caller to provide statistics. 
Note that the niumber of messages does not match with the number of
monitor point and values sent to the core of the IAS because the
publisher groups more values in the same message.