Allows to start/stop the sending of monitor point values to the core of the IAS.
startSending() and stopSending() allows to start and stop the sending of
monitor point values to the core of the IAS. When the sending is
inhibited, the values to be sent to the core are lost forever.

In addition, the setUp() and tearDown() methods to allocate and free
resource have been added to the MonitorPointSender interface. It
triggered some renaming in the PublisherBase where such methods already
existed.

Provides tests to check the new functionality.