Introduced google guice for dependency injection and provides few publisher implementations
There are 2 publisher implementations: one that write data in a passed
text file and one that forward events to a listener. Both mostly
intended for testing.
The one that forward events to a listener was, in fact, in the test
folder.

MonitoredSystemData has a method to produce a JSON string (not yet
tested).

Note that dependency injection has not been tested yet.