Limits the number of monitor point values t send in a time interval.
The size of the buffer (the) can be limited to avoid memory overflows by
setting the org.eso.ias.plugin.buffersize java property. If the size of
the buffer grows over the limit, the monitor points are immediately sent
to the core of the IAS.

The PublisherBase has been renamed to BufferedPublisherBase to stress
that monitor ponts are buffered before being sent.