Adds the initial version of the KafkaPublisher
The KafkaPublisher publishes monitor point events in a KafkaTopic. It is
not yet complete as partition number, topic name and connection
properties are all hardcoded.
It also adds the initial version of the tests, KafkaPublisherTest that
uses a SimpleKafkaConsumer to check if published events are really
published.

Both the consumer and the test need improvements but events ar pblished
and received by the consumer.