Builds java and scala API documentation (python still missing)
iasBuildApiDocs build the java and scala API documentation in the passed
folder. Scala API docs in a scala subdirectory and java API docs in a
java subfolder.
The script delegates the execution of javadoc and scaladoc to the
objects in the IASApiDocs python module.
Scala documentation includes also java as it can be more convenient to
have all the documentation under the same web page.
Generation of python API documentation still missing.