Improves the definition of the ConverterStream and cleans up unused classes
The id of the converter and the function to map input to output strings
are part of the base class while specific parameter have to be passed in
the constructor of the specialised class.

With the stream the code is much simpler and several classes are not
needed anymore. They have been removed, togeter with their tests.