Sets auto.offset.reset props to earliest to avoid loosing records.
Without this setting the test was sometimes loosing records. It happened
at the beginning when the topic was not used for long time probably
because kafka rearrange in the group internally. This setting allows to
receive all the records published since the last committed reading
ensuring to read all the records even those published before kafka is
really delivering data to consumers.
Note that there is the risk to read the same data more then once if the
previous run did not commit properly (for example a crash of the
process).
Processing the data twice or miss some data is not a problem because of
the functioning of the IAS.