Adds a test to ensure the translation works even in case of errors
The errors tested here are those about an unknown moitor point i.e. one
that is not found in the CDB and for which it is not possible to
understand the type.
There could be other errors at run-time like for example a parsing error
that are not tested here.