Assigns ASCEs to levels
Assignment of ASCEs to level is a fundamental step for calculating the
output of the DASU, produced by one of its ASCEs.
Inputs coming from IASIOs queues are the input of ASCEs at level 0 that
are the ASCEs that depends only on external IASIOs.
ASCEs at level 1 are those that requires the IASIOs coming from the
queues and those generated by the ASCEs at level 0.
In general ASCEs at level l requires the input coming from IASIO queues
pluse the IASOIOs generated by the ASCEs belonging to levels [0,l-1].