Removes a wrong assert
The size of the inputs is not necessarily equal to the size of the
required inputs.
In fact the beginning the ASCE has 0 inputs, inputs that  will be passed
by the DASU as soon as it receives them. At run time the max number of
inputs must be equal to the size of the required inputs.