Improve the check of validity for parent identifier
@see Issue #26
In particular thie check allows to instantiate IASIO identifiers with
None parents that is the case of the ACSE that initially knows the ID of
the IASIOs it expects in input but does not know who produces them
before they arrives. This is fine as the computation of the TF is based
only on the IDs and not on the ID of the parents.