The constructor of the ASCE takes a set of IASIOinsted of their IDs
The ASCE needs the IASIOs to be able to update them From the IDs it is
not possible to build the IASIOs (even with a None value) because their
type is unknown.
The IASIOs passed in the constructor have aNone value and their
identifiers have no parent because the ASCE does not know yet who
produces them but only their ID. When the DASU pass the inputs to run
the TF of the ASCE then the IASIOs are also updated with the parent
identifier.
Note that the parent ID can change at run time if the IASIO passes from
a different converter or a DASU is redeployed in another Supervisor.
This is perfectly legal as the running ID is mostly used for debugging
and logging.