Improves the way of getting the core size from java properties
The class has an auxiliary constructor that takes the size of the pool
from the java properties or, if not found, uses a default based on the
number of available cores.

When in future we will have more DASUs running in a supervisor. the
scheduled executor must be provided by the Supervisor as it is shared
between all the DASUs runnin in the same Supervisor.