Adds the InputsUndefined state.
InputsUndefined is the state reached by the ASCE after the
initialization when some of its inputs have not yet been received and
have a null/empty value. In that case the execution of the TF is
inhibited and the ASCE does not produce any output.
The event InputsInitialized() have been added to signal that all the
inputs have been initialized so that the ASCE can change its state and
run the TF.