update(..) returns a value of (Option[InOut[_]], AsceStates.State)
When the ASCE has not yet been initialized and some of its input have an
empty value, it cannot produce an output but returns a None.
This is a temporary situation: when all the inputs arrive then the ASCE
produces a voule running the TF even if the inputs have been update long
before. In that case the validity reports the problem.
@see [Issue
#37](https://github.com/IntegratedAlarmSystem-Group/ias/issues/37)