Refreshes the output at regular time intervals
The automatic refresh of the output is triggered by a scheduled task
that must be activated by calling enableAutoRefreshOfOutput(). It can be
disabled by calling disableAutoRefreshOfOutput().

The sending production of the output happens immediately if the timer
task is disabled otherwise can be delayed if the update is requested too
often (i.e before the minAllowedRefreshRate elapses or when the timer
task will be executed before the minAllowedRefreshRate. In both cases,
the inputs are stored in a buffer implemented as a map to keep only the
last value if the same input is received many times.