Removes the InOutValue move timestamp and value in InOut. Reviews equals
There is no need to consider 2 InoUt equals for all their fields but the
timestamp. I f it is the case then we will introduce a new method that
make such comparison more evident.
equals (and the test about equality) has been removed because scala case
classes define equals and hashCode for us