Handling of the Validity in the InOut reviewd (@see Issue #52)
The validity must always be calculated because it depends on
- the actual timestamp against the last update time and the refresh rate
- the min validity of the IASIOs in input to the TF that generates a
IASIOs, if the IASIOs is generated by a ASCE.

The validity is not a property but returned by a method that makes the
calculation.

The IASValue published in the BSDB will have the validity calculated at
the moment of the conversion from a IASIOs. However, before using the
value, its validity must again be validated against the actual
timestamp.