Sets the oprational mode of the output by injecting the one of its input.
The TF is in charge to set the value AND the mode of the output.
In the case of only one input, it makes sense to have the same mode in the output.

But when the TF is dealing with more inputs then the mode should be set
in agreement with the mode of the inputs and other factors.