Distinguish the validity from the inputs and that of the output
A IASIO can either be the input of a ASCE or the output. In the former
case the validity depends from the validity received from the BSDB (i.e.
the IASValue) in the latter the validity depends on the validity of the
inputs.
Onl;y one of the 2 validity must be defined and the other must be empty.