New version of the table #80
This version separate the deployment of the DASUs in the supervisor
because of the templated DASUs.
With templates the relation between DASU and Supervisor is not OneToMany
as it was before.

I preferred to link ASCEs to DASUs that are both templated breaking the
links to the supervisor. To keep such relation we should store also the
instances and link [template, instance number] to Supervisor.
I have instead separated the deployment in the DASu from the definition
of DASUs and ASCEs. This way the deployment can be set independently of
DASUs definitions.

This new version has pros and cons. In particular the deployment can be
set to run DASUs that do not exists and this condition is not checked
while inserting data in RDB but will be shown when starting a container.
We can develop tools to check RDB consistency at interfaces to
manipulate the RDB that check this and other conditions.