Adds a get method that return a value being its identifier templated or not
This method is very useful for TF running on templated ASCE as the
instance is transparently handled by the method.
In this way the user does not need to know much of the internal details
of the core like the string pattern used for templated identifiers,
neither the API of the IDentifier to build a concrete ID from its
template and the number of the instance.
@see #80