Generate a new segment every 5 minutes and save only 1 hour of logs
kafka keeps only a given interval of time that we set to 1 hour. However
the deletion is done only on segments tha have no consumer.
The default is ti create a new segment every week so to really have
segments to delete we mode the generation of a new segment every 5
minutes
@see #94