Adds a field to set constraints calculating the validity #103
The constraint is about limiting the calculation of the validity to a
give set of inputs instead of all of them.
The constraint is passed by setting a set of IDs of the inputs to
consider discarding the others.
This field will be used by the ASCE and the DASU when assessing the
validity