New algorithm for the validity #103
The ASCEs send the output with the validity calculated depending on the
validity of the inputs and their timestamps.
The DASU must then only check the validity set in the last produced
IASValue and the production time against the current timestamp. There is
no need to check the validity of the inputs from the ASCE as this step
is transparently done by the ASCEs after running the TF.

The check of when the value must be published, mustSendOutput, takes
into account also changes in the properties and in the list of
dependents.