Accepts only inputs whose timestamp is newer than the input in the map #39
inputsReceived delivers a list of inputs to process and it can happen that the same input appears more then ones in the list. In that case we must be sure to get only the newest input and discard all the others.
We defined a filter, acceptIasValue, that accepts only values whose ID is the ID of one of th einput of the ASCE AND whose production timestamp is newer that the timestamp of the input to process stored in the map.