Adds a test to ensure that a supervisor restarts when the TF changed
The test starts a supervisor and checks the output it produces.
Later it changes the TF and send a broadcast TF_CHANGED.
The supervisor, upon receiving the comamnd, restarts and gets the new TF definition.
The test sends another input to check if the value produced by the supervisor is produced by the new TF