The CommandManager does not need the Identifier
As there are tools that have no fullRunningID, to let this tool work we get rid of the Identifier and use only the id.
This is perfectly fine as soon as there are no IAS tools with the same id.

The problem of the Identifer (and the fullRunningId appeared already for the HBs that, in fact, uses a HB instead of a IdentifierType.

To be 100% in the safe side, the CDB checker should also report if there are tools with the same ID.