Reviews the KafkaStringConsumer
The ConsumerRebalanceListener is implemented by the KafkaStringsConsumer class

The polling CountDownLatch has been removed.

isReady() can be used to know when the consumer is ready to get events (i.e. initialized and connected to at list one partition of the topic).

startGettingEvents() setup the consumer to get events but it requires some time before it really start getting the events (it must get a partition befor and it can cause a rebalance).