Waits until subscribed to the partition
There is a bug in the confluent kafka consumer that does not create
a topic if it does not exist.
The KafkaValueConsumer checks if subscribed to the topic when getting
a message. The topic will be created by someone else and at that point
in time the consumer will get the messages.
Before this commit the consumer exited with error.